+++ backend/src/routes/donations.js
@@ -1,5 +1,5 @@
 import express from "express";
-import {getDonation, getDonations, registerDonation, deleteDonation} from "../controllers/donation.js"
+import {getDonation, getDonations, registerDonation, deleteDonation} from "../controllers/donations.js"
 
 const router = express.Router();
